+++ src/AnalysisG/modules/tools/include/tools/tools.h
@@ -30,7 +30,7 @@ class tools
         bool has_value(std::vector<std::string>* data, std::string trg); 
 
         std::vector<std::string> split(std::string in, std::string delim);
-        std::vector<std::string> split(std::string in, int n);
+        std::vector<std::string> split(std::string in, size_t n);
         std::string hash(std::string input, int len = 18);
         std::string lower(std::string*); 
 
@@ -43,11 +43,11 @@ class tools
         // template functions
         template <typename G>
         std::vector<std::vector<G>> discretize(std::vector<G>* v, int N){
-            int n = v -> size(); 
+            size_t n = v -> size(); 
             typename std::vector<std::vector<G>> out; 
             out.reserve(int(v -> size()/N)); 
-            for (int ib = 0; ib < n; ib += N){
-                int end = ib + N; 
+            for (size_t ib = 0; ib < n; ib += N){
+                size_t end = ib + N; 
                 if (end > n){ end = n; }
                 out.push_back(std::vector<G>(v -> begin() + ib, v -> begin() + end)); 
             }
@@ -57,7 +57,7 @@ class tools
         template <typename g>
         g max(std::vector<g>* inpt){
             g ix = inpt -> at(0); 
-            for (int t(1); t < inpt -> size(); ++t){
+            for (size_t t(1); t < inpt -> size(); ++t){
                 if (inpt -> at(t) <= ix){continue;}
                 ix = inpt -> at(t); 
             }
@@ -67,7 +67,7 @@ class tools
         template <typename g>
         g min(std::vector<g>* inpt){
             g ix = inpt -> at(0); 
-            for (int t(1); t < inpt -> size(); ++t){
+            for (size_t t(1); t < inpt -> size(); ++t){
                 if (inpt -> at(t) >= ix){continue;}
                 ix = inpt -> at(t); 
             }
@@ -77,7 +77,7 @@ class tools
         template <typename g>
         g sum(std::vector<g>* inpt){
             g ix = 0; 
-            for (int t(0); t < inpt -> size(); ++t){ix += (*inpt)[t];}
+            for (size_t t(0); t < inpt -> size(); ++t){ix += (*inpt)[t];}
             return ix; 
         }
